+++ test/test_jobs.rb
@@ -27,9 +27,9 @@ class JobsTestCase < TestCaseWithSplunkConnection
   # There is a convenience method on service to create an asynchronous
   # search job. Test it the same way.
   #
-  def service_create_and_idempotent_cancel
+  def test_service_create_and_idempotent_cancel
     jobs = @service.jobs
-    job = @service.create(QUERY)
+    job = @service.create_search(QUERY)
     assert_true(jobs.has_key?(job.sid))
     job.cancel()
     assert_eventually_true() { !jobs.has_key?(job.sid) }
@@ -100,6 +100,77 @@ class JobsTestCase < TestCaseWithSplunkConnection
     stream = @service.create_export("search index=_internal earliest=-2d | stats count(_raw) by method")
     results = ResultsReader.new(stream).to_a()
     assert_true(3 >= results.length())
+  end
+
+  ##
+  # Test that oneshot jobs have no <sg> elements in the XML they return
+  # by default.
+  #
+  def test_oneshot_has_no_segmentation_by_default
+    omit_if(@service.splunk_version[0] == 4)
+    stream = @service.create_oneshot("search index=_internal GET | head 3")
+    assert_false(stream.include?("<sg"))
+  end
+
+  ##
+  # Are <sg> elements returned in the XML from a oneshot job when we pass
+  # the option segmentation=raw?
+  #
+  def test_oneshot_has_forced_segmentation
+    omit_if(@service.splunk_version[0] == 4)
+    stream = @service.create_oneshot("search index=_internal GET | head 3",
+                                     :segmentation => "raw")
+    assert_true(stream.include?("<sg"))
+  end
+
+  ##
+  # Test that export jobs have no <sg> elements in the XML they return by
+  # default.
+  #
+  def test_export_has_no_segmentation_by_default
+    omit_if(@service.splunk_version[0] == 4)
+    stream = @service.create_export("search index=_internal GET | head 3")
+    assert_false(stream.include?("<sg"))
+  end
+
+  ##
+  # Export jobs should have <sg> elements in the XML they return when a
+  # value is passed to the segmentation argument to make it so.
+  #
+  def test_export_has_forced_segmentation
+    omit_if(@service.splunk_version[0] == 4)
+
+    stream = @service.create_export("search index=_internal GET | head 3",
+                                     :segmentation => "raw")
+    assert_true(stream.include?("<sg"))
+  end
+
+  ##
+  # Results and preview on a search job should have no segmentation
+  # by default.
+  #
+  def test_asynchronous_job_has_no_segmentation_by_default
+    omit_if(@service.splunk_version[0] == 4)
+    job = @service.jobs.create("search index=_internal GET | head 3")
+    until job.is_done?()
+      sleep(0.1)
+    end
+    assert_false(job.events().include?("<sg"))
+    assert_false(job.preview().include?("<sg"))
+  end
+
+  ##
+  # Results and preview on a search job should have segmentation when
+  # it is forced.
+  #
+  def test_asynchronous_job_has_segmentation_when_forced
+    omit_if(@service.splunk_version[0] == 4)
+    job = @service.jobs.create("search index=_internal GET | head 3")
+    until job.is_done?()
+      sleep(0.1)
+    end
+    assert_true(job.events(:segmentation => "raw").include?("<sg"))
+    assert_true(job.preview(:segmentation => "raw").include?("<sg"))
   end
 
   def test_each_and_values